Cached space api file used on update failure. First recheck after 2 min.
If the failure counter is 0 the update schedule is set to the smallest
defined schedule and with every next failure the schedule will be rotated.
This means that on the second failure it will take 5min until it's rechecked.
The schedule rotator will stop when the longest schedule has been reached
or if the space server has replied.