Seperate json files for each specs version. apienv.js is now generated in php.
In order to use a php json schema validator there's a need to have a seperate
file for each specs version. The router got extended by the apienv and specs
delegators. When using the apienv delegator the javascript for JSV is returned
and the actual spaceapi specs when using the specs delegator.

Mapping:

   <openspacelint>/c/js/apienv.js => apienv delegator
   <openspacelint>/specs/<version> => specs delegator

Examples:

   <openspacelint>/specs/0.7 returns a (json) message that this version is undefined
   <openspacelint>/specs/0.8 returns version 0.8 of the space api specs
   <openspacelint>/specs/0.9 returns version 0.9 of the space api specs